move test file out of src and update npm run test script
