+++ package.json
@@ -62,10 +62,10 @@
     "package": "webpack --mode production --devtool hidden-source-map",
     "compile-ts": "tsc -p . --outDir out",
     "compile-antlr": "antlr4ts -visitor -o src/generated/ src/grammar/VisualCobol.g4",
-    "watch-tests": "tsc -p . -w --outDir out",
-    "pretest": "npm run compile-tests && npm run compile && npm run lint",
+    "watch-tests": "tsc -p ./tsconfig.test.json -w --outDir out",
+    "compile-tests": "tsc -p ./tsconfig.test.json --outDir out",
     "lint": "eslint src",
-    "test": "vscode-test"
+    "test": "npm run compile-tests && vscode-test"
   },
   "repository": {
     "type": "git",
